requirements.txt
pyserial==3.5
wheel==0.36.2
construct==2.10.67